+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Settings page. Stored in cloud with cache in local storage.
   - Auto skip time. Automatically skip the exercise after a certain time.
   - WPM Decay. Used to sword words on slowest WPM.
+- Words page. Shows all the practice words, their WPM and last practiced time.
 
 ## 2024-07-06 - 2024-07-31